Move main function to main.go
