+++ main.go
@@ -1,0 +1,63 @@
+package main
+
+import (
+	"log"
+	"net/http"
+	"os"
+
+	"github.com/alecthomas/kingpin"
+	"github.com/prometheus/client_golang/prometheus"
+)
+
+func main() {
+	var (
+		app                                           = kingpin.New("postfix_exporter", "Prometheus metrics exporter for postfix")
+		listenAddress                                 = app.Flag("web.listen-address", "Address to listen on for web interface and telemetry.").Default(":9154").String()
+		metricsPath                                   = app.Flag("web.telemetry-path", "Path under which to expose metrics.").Default("/metrics").String()
+		postfixShowqPath                              = app.Flag("postfix.showq_path", "Path at which Postfix places its showq socket.").Default("/var/spool/postfix/public/showq").String()
+		postfixLogfilePath                            = app.Flag("postfix.logfile_path", "Path where Postfix writes log entries. This file will be truncated by this exporter.").Default("/var/log/postfix_exporter_input.log").String()
+		systemdEnable                                 bool
+		systemdUnit, systemdSlice, systemdJournalPath string
+	)
+	systemdFlags(&systemdEnable, &systemdUnit, &systemdSlice, &systemdJournalPath, app)
+
+	kingpin.MustParse(app.Parse(os.Args[1:]))
+
+	var journal *Journal
+	if systemdEnable {
+		var err error
+		journal, err = NewJournal(systemdUnit, systemdSlice, systemdJournalPath)
+		if err != nil {
+			log.Fatalf("Error opening systemd journal: %s", err)
+		}
+		defer journal.Close()
+	}
+
+	exporter, err := NewPostfixExporter(
+		*postfixShowqPath,
+		*postfixLogfilePath,
+		journal,
+	)
+	if err != nil {
+		log.Fatalf("Failed to create PostfixExporter: %s", err)
+	}
+	prometheus.MustRegister(exporter)
+
+	http.Handle(*metricsPath, prometheus.Handler())
+	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
+		_, err = w.Write([]byte(`
+			<html>
+			<head><title>Postfix Exporter</title></head>
+			<body>
+			<h1>Postfix Exporter</h1>
+			<p><a href='` + *metricsPath + `'>Metrics</a></p>
+			</body>
+			</html>`))
+		if err != nil {
+			panic(err)
+		}
+	})
+
+	log.Print("Listening on ", *listenAddress)
+	log.Fatal(http.ListenAndServe(*listenAddress, nil))
+}
